TASK-2 and TASK-4: As a user I can mark a todo as completed. As a user I can see a confirmation message when I add/delete a todo.
I am including two tasks on this commit because part of TASK-4 had been done previously, so only thing missing was to alert user on Todo creation, which is completed now.
Added completed field to model and created migration. The migration was created using Flask-Migrate library, with some tweaks manually made. Also renamed initial migration to a more descriptive name.
Implemented complete todo funcionality. Changed the todo_delete method for a todo_update method and in the form for delete todo included a hidden tag _method that will be checked in the update method. This is not ideal but HTML does not support all HTTP methods, so this is a way to simulate a DELETE request. For a real life application, a better way would be to create a Flask Middleware that would change the request method based on the form tag and have different methods for delete and update.
Fixed tests delete_todo functionality to comply with new route params, now the test fixture sends the expected _method param as part of the request.
Using macros to refactor multiple and single todo view functionality. Functionality for mark todo as completed and remove a todo was repeated in the todo list and single todo view. Refactored the code using Jinja macros.
Added test for completeTodo functionality.

+++ alayatodo/views.py
@@ -94,8 +94,16 @@ def todos_POST():
 
 @app.route('/todo/<id>', methods=['POST'])
 @require_login
-def todo_delete(id):
+def todo_update(id):
     todo = db.session.query(Todo).filter(Todo.id == id, Todo.user_id == session['user_id']).first_or_404()
-    db.session.delete(todo)
+    if request.form.get('_method', '').upper() == 'DELETE':
+        flash('Todo has been deleted.', 'danger')
+        db.session.delete(todo)
+    else:
+        completed = request.form.get('completed') is not None
+        todo.completed = completed
+        flash('Todo has been marked as {}completed.'.format('' if
+                                                            completed else 'not '), 'success')
+        db.session.add(todo)
     db.session.commit()
     return redirect('/todo')
